+++ src/main/java/jwp/controller/LogOutController.java
@@ -1,18 +1,19 @@
 package jwp.controller;
 
-import core.mvc.Controller;
-
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import java.io.IOException;
 
-
-public class LogOutController implements Controller {
-
+@WebServlet("/user/logout")
+public class LogOutController extends HttpServlet {
     @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         HttpSession session = req.getSession();
         session.removeAttribute("user");
-        return "redirect:/";
+        resp.sendRedirect("/");
     }
 }
